ref: moved create-question dialog into separate component
